fix cat << eof << eof

+++ src/pipes/input.c
@@ -1,42 +1,41 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   redirection01.c                                    :+:      :+:    :+:   */
+/*   input.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: igvisera <igvisera@student.42madrid.com>   +#+  +:+       +#+        */
+/*   By: igvisera <igvisera@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/02 22:03:53 by igvisera          #+#    #+#             */
-/*   Updated: 2025/03/10 17:35:01 by igvisera         ###   ########.fr       */
+/*   Updated: 2025/03/12 18:32:23 by igvisera         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/minishell.h"
 
-int	open_input_files(t_token *data)
+int open_input_files(t_token *data)
 {
-	int				fd;
-	t_redirect_file	*infile;
-	t_list			*infiles;
+    int         fd = -1;
+    t_list      *current = data->infiles;
+    t_redirect_file *infile = NULL;
 
-	fd = -1;
-	infiles = data->infiles;
-	while (infiles)
-	{
-		infile = (t_redirect_file *)infiles->content;
-		fd = open(infile->value, O_RDONLY);
-		if (fd < 0)
-		{
-			g_exit_status = 1;
-			perror("open input");
-			return (-1);
-		}
-		if (ft_strncmp(infile->value, ".heredoc.tmp", 12) == 0)
-			unlink(infile->value);
-		if (infiles->next)
-			close(fd);
-		infiles = infiles->next;
-	}
-	return (fd);
+	while (current)
+    {
+        infile = (t_redirect_file *)current->content;
+        int temp_fd = open(infile->value, O_RDONLY);
+        if (temp_fd < 0)
+        {
+            g_exit_status = 1;
+            perror("open input");
+            return -1;
+        }
+        if (fd != -1)
+            close(fd);
+        fd = temp_fd;
+        current = current->next;
+    }
+    if (infile && ft_strncmp(infile->value, ".heredoc.tmp", 12) == 0)
+        unlink(infile->value);
+    return fd;
 }
 
 int	redirect_input(t_ast *ast)
